Separate multer configuration into its own module
- Moved Multer configuration to a dedicated config file.
- Created a MulterConfigService to handle dynamic configuration using ConfigService.
- Updated AuthModule to register MulterModule with the new configuration.
- Refactored AuthWorkerController to use the new Multer configuration setup.
